+++ back/src/main/java/com/seguridadlimite/models/permiso/domain/PermisoTrabajoAlturasBasico.java
@@ -37,7 +37,7 @@ public class PermisoTrabajoAlturasBasico {
     private String horaFinal;
 
     @Column(name = "id_nivel")
-    private Long idNivel;
+    private Integer idNivel;
 
     @JoinColumn(name = "id_nivel", referencedColumnName = "id", insertable = false, updatable = false)
     @ManyToOne(fetch = FetchType.EAGER)
